+++ apps/tk-capture-book/app/(auth)/(tabs)/frame.tsx
@@ -1,0 +1,10 @@
+import { View } from "react-native";
+import { Text } from "@/components/ui/text";
+
+export default function FrameScreen() {
+  return (
+    <View className="flex-1">
+      <Text>Frame</Text>
+    </View>
+  );
+}
