+++ apps/tk-capture-book/consts/locales/en.ts
@@ -3,6 +3,9 @@ export const en = {
     camera: {
       tabTitle: "Camera",
     },
+    frame: {
+      tabTitle: "Frame",
+    },
     login: {
       title: "Welcome",
       subtitle: "Login to your account",
