+++ apps/tk-capture-book/features/book/components/book-search-input.tsx
@@ -1,0 +1,34 @@
+import { Ionicons } from "@expo/vector-icons";
+import React from "react";
+import { Pressable, TextInput, View } from "react-native";
+
+interface BookSearchInputProps {
+  searchText: string;
+  setSearchText: (text: string) => void;
+  placeholder?: string;
+}
+
+export const BookSearchInput: React.FC<BookSearchInputProps> = ({
+  searchText,
+  setSearchText,
+  placeholder = "검색어를 입력하세요",
+}) => {
+  return (
+    <View className="relative flex-row items-center">
+      <TextInput
+        className="flex-1 p-2 pr-8 border border-gray-300 rounded-lg"
+        placeholder={placeholder}
+        value={searchText}
+        onChangeText={setSearchText}
+      />
+      {searchText.length > 0 && (
+        <Pressable
+          className="absolute right-2"
+          onPress={() => setSearchText("")}
+        >
+          <Ionicons name="close-circle" size={20} color="#9CA3AF" />
+        </Pressable>
+      )}
+    </View>
+  );
+};
